+++ CHANGELOG.md
@@ -7,9 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Changed
-- Configuration format for whitelisting users has been changed to make it easier to 
-  add further global filters  
+- Configuration format for whitelisting users has been changed to make it easier to
+  add further global filters
 - enable and disable filters via the config
+- output the number of filtered pull requests
 
 ## [0.6.0] - 2017-10-13